Initial PPO training script, inference refactoring.
