+++ src/apps/infer/cleanrl_ppo_infer.py
@@ -1,58 +1,23 @@
 import gymnasium as gym
-import numpy as np
 import torch
-from torch.distributions import Categorical
-import torch.nn as nn
 import typer
 
-
-def layer_init(layer: nn.Module, std: float = np.sqrt(2), bias_const: float = 0.0) -> nn.Module:
-    torch.nn.init.orthogonal_(layer.weight, std)
-    torch.nn.init.constant_(layer.bias, bias_const)
-    return layer
+from apps.train.ppo_train import Agent
 
 
-class Agent(nn.Module):
-    def __init__(self, single_observation_space_shape: tuple[int], single_action_space_shape: np.ndarray) -> None:
-        super().__init__()
-        self.critic = nn.Sequential(
-            layer_init(nn.Linear(np.array(single_observation_space_shape).prod(), 64)),
-            nn.Tanh(),
-            layer_init(nn.Linear(64, 64)),
-            nn.Tanh(),
-            layer_init(nn.Linear(64, 1), std=1.0),
-        )
-        self.actor = nn.Sequential(
-            layer_init(nn.Linear(np.array(single_observation_space_shape).prod(), 64)),
-            nn.Tanh(),
-            layer_init(nn.Linear(64, 64)),
-            nn.Tanh(),
-            layer_init(nn.Linear(64, single_action_space_shape), std=0.01),
-        )
-
-    def get_value(self, x: torch.Tensor) -> torch.Tensor:
-        return self.critic(x)
-
-    def get_action_and_value(
-        self, x: torch.Tensor, action: torch.Tensor = None
-    ) -> tuple[torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor]:
-        logits = self.actor(x)
-        probs = Categorical(logits=logits)
-        if action is None:
-            action = probs.sample()
-        return action, probs.log_prob(action), probs.entropy(), self.critic(x)
-
-
-def main(agent_path: str, steps_count: int = 1000) -> None:
-    seed = 0
-
-    env = gym.make("Acrobot-v1", render_mode="human")
+def main(
+    agent_path: str,
+    env_id: str = "Acrobot-v1",  # Environment ID
+    total_timesteps: int = 1000,  # Total number of timesteps
+    seed: int = 0,  # Random seed
+) -> None:
+    env = gym.make(env_id, render_mode="human")
     observation, _ = env.reset(seed=seed)
 
     agent = Agent(env.observation_space.shape, env.action_space.n)
     agent.load_state_dict(torch.load(agent_path))
 
-    for _ in range(steps_count):
+    for _ in range(total_timesteps):
         action, _, _, _ = agent.get_action_and_value(torch.Tensor(observation))
         observation, _, terminated, truncated, _ = env.step(action)
